Move logic inside TanksOverview and remove Tanks component
